[rebroadcast] Introduce a mempool rebroadcast module
As the first step of extracting the rebroadcast logic from the wallet to the
node, introduce a standalone module that calculates the set of rebroadcast
candidates at the time. This is not yet used to actually send the transaction
candidates to peers.

One fundamental difference with the existing wallet rebroadcast logic is that
we apply the logic to all transactions, not just "mine". In order to prevent
spam, we need to carefully select which transactions to rebroadcast. As of this
commit, we select the transactions in the mempool that have the highest
feerate.