[p2p] Periodically rebroadcast transactions to peers
Rebroadcasts are scheduled on a per-peer timer with added randomness to avoid
any privacy leaks or network synchronization issues. When the peer's timer
pops, we attempt to rebroadcast the transactions selected by the txrebroadcast
module. There are two main filters that determine if this changes any behavior:
transactions will be de-duplicated before adding to `setInventoryTxToSend`, and
they will later pass through the `filterInventoryKnown` bloom filter for the
peer before being announced over the connection.