[rebroadcast] Periodically calculate and cache minimum fee rate.
As of this commit, we regularly calculate and update our cached values around
values such as the minimum fee rate for a transaction to be considered at the
top of the mempooland the chain tip at the time the cache ran. We do not yet
use them to filter the transactions that are rebroadcasted.