[net_processing] Defer initializing m_addr_known
Use SetupAddressRelay to only initialize `m_addr_known` when a peer sends us an
addr related message, unless they are an outbound block-relay-only peer.

We update the definition of RelayAddrsWithConn to check for presence of
`m_addr_known`. This should be a strict superset of the previous condition (not
block-relay-only peer), since SetupAddressRelay ensures that we will never
initailize the bloom filter for block-relay-only connections.

These changes intend to mitigate address blackholes. Since an inbound peer has
to send us an addr related message to become eligible as a candidate for addr
relay, this should reduce our likelihood of sending them self-announcements.