[net_processing] Defer initializing m_addr_known
Use SetupAddressRelay to only initialize `m_addr_known` as needed. For outbound
peers, we initialize the filter before sending our self announcement (not
applicable for block-relay-only connections). For inbound peers, we initialize
the filter when we get an addr related message (ADDR, ADDRV2, GETADDR,
SENDADDRV2).

We update the definition of RelayAddrsWithConn to check for presence of
`m_addr_known`. This should be a strict superset of the previous condition (not
block-relay-only peer), since SetupAddressRelay ensures that we will never
initailize the bloom filter for block-relay-only connections.

These changes intend to mitigate address blackholes. Since an inbound peer has
to send us an addr related message to become eligible as a candidate for addr
relay, this should reduce our likelihood of sending them self-announcements.