[test] Test that we only relay addrs to peers who sent us addr related messages
The `AddrReceiver` class used in the existing tests sends a `getaddr` message
to indicate that it is interested in participating in address relay. This
commit introduces another class, `AddrBlackhole` which does not initiate any
addr related messages. The tests check that addresses received via the network
are forwarded to the `AddrReceiver` and not the `AddrBlackhole`